+++ macunpack/bin.c
@@ -14,7 +14,7 @@ void
 bin (char *header, int data_size, int UMcp)
 {
     char hdr[INFOBYTES];
-    unsigned long rsrcLength, dataLength;
+    uint32_t rsrcLength, dataLength;
 
     hdr[0] = getb(infp);
     (void)ungetc(hdr[0], infp);
@@ -38,7 +38,7 @@ bin (char *header, int data_size, int UMcp)
 #ifdef SCAN
 	do_idf("", COPY);
 #endif /* SCAN */
-	mcb(header, (unsigned long)in_data_size, (unsigned long)in_rsrc_size,
+	mcb(header, (uint32_t)in_data_size, (uint32_t)in_rsrc_size,
 	    in_ds + in_rs);
 	ds_skip = 0;
 	rs_skip = 0;
