+++ macunpack/de_compress.c
@@ -19,20 +19,20 @@
 
 static int n_bits;				/* number of bits/code */
 static int maxbits;			/* user settable max # bits/code */
-static long maxcode;			/* maximum code, given n_bits */
-static long maxmaxcode;			/* should NEVER generate this code */
+static int32_t maxcode;			/* maximum code, given n_bits */
+static int32_t maxmaxcode;			/* should NEVER generate this code */
 # define MAXCODE(n_bits)	((1 << (n_bits)) - 1)
 
-static long htab [HSIZE];
+static int32_t htab [HSIZE];
 static unsigned short codetab [HSIZE];
 
 #define tab_prefixof(i) codetab[i]
 #define tab_suffixof(i)	((unsigned char *)(htab))[i]
 #define de_stack	((unsigned char *)&tab_suffixof(1<<BITS))
 
-static long free_ent = 0;			/* first unused entry */
+static int32_t free_ent = 0;			/* first unused entry */
 
-static long getcode();
+static int32_t getcode();
 
 static int clear_flg = 0;
 
@@ -45,11 +45,11 @@ static int clear_flg = 0;
 
 static int toread;
 
-void de_compress(unsigned long ibytes, int mb)
+void de_compress(uint32_t ibytes, int mb)
 {
     register unsigned char *stackp;
     register int finchar;
-    register long code, oldcode, incode;
+    register int32_t code, oldcode, incode;
 
     toread = ibytes;
     maxbits = mb;
@@ -124,10 +124,10 @@ static char *core_ptr;
 static int file_bytes();
 static int core_bytes();
 
-static long 
+static int32_t 
 getcode (void)
 {
-    register long code;
+    register int32_t code;
     static int offset = 0, size = 0;
     static unsigned char buf[BITS];
     register int r_off, bits;
